3DS Max: refactoring
